+++ src/GUI/MapLoader.cpp
@@ -5,7 +5,7 @@
 #include <QJsonObject>
 #include <QJsonValue>
 
-const std::map<std::string, std::vector<std::map<std::string, int>>>
+const std::map<std::string, std::vector<std::map<std::string, double>>>
 MapLoader::loadMap(QString path) {
   QFile file(path);
   if (!file.open(QIODevice::ReadOnly)) {
@@ -15,16 +15,16 @@ MapLoader::loadMap(QString path) {
   QJsonDocument doc = QJsonDocument::fromJson(file.readAll());
   QJsonObject obj = doc.object();
 
-  std::map<std::string, std::vector<std::map<std::string, int>>> map;
+  std::map<std::string, std::vector<std::map<std::string, double>>> map;
 
   for (const auto& key : obj.keys()) {
     QJsonValue value = obj.value(key);
     if (value.isArray()) {
-      std::vector<std::map<std::string, int>> vec;
+      std::vector<std::map<std::string, double>> vec;
       for (const auto& item : value.toArray()) {
-        std::map<std::string, int> m;
+        std::map<std::string, double> m;
         for (const auto& k : item.toObject().keys()) {
-          m[k.toStdString()] = item.toObject().value(k).toInt();
+          m[k.toStdString()] = item.toObject().value(k).toDouble();
         }
         vec.push_back(m);
       }
